+++ platform-operator/controllers/verrazzano/component/certmanager/certmanager_test.go
@@ -1,12 +1,20 @@
-// Copyright (c) 2021, 2022, Oracle and/or its affiliates.
+// Copyright (c) 2021, 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
 package certmanager
 
 import (
 	"context"
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
 	"fmt"
+	"math/big"
+	"net"
 	"testing"
+	"time"
 
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
 
@@ -17,12 +25,10 @@ import (
 	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
-	vzconst "github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
@@ -31,7 +37,6 @@ import (
 	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	clipkg "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
 const (
@@ -355,177 +360,6 @@ func TestIsCABothPopulated(t *testing.T) {
 	assert.Error(t, err)
 }
 
-// TestCreateCAResources tests the createOrUpdateCAResources function.
-func TestCreateCAResources(t *testing.T) {
-	// GIVEN that a secret with the cluster CA certificate does not exist
-	// WHEN a call is made to create the CA resources
-	// THEN the call succeeds and an Issuer, Certificate, and ClusterIssuer have been created
-	localvz := defaultVZConfig.DeepCopy()
-	localvz.Spec.Components.CertManager.Certificate.CA = ca
-
-	client := fake.NewClientBuilder().WithScheme(testScheme).Build()
-
-	opResult, err := createOrUpdateCAResources(spi.NewFakeContext(client, localvz, nil, false, profileDir))
-	assert.NoError(t, err)
-	assert.Equal(t, controllerutil.OperationResultCreated, opResult)
-
-	// validate that the Issuer, Certificate, and ClusterIssuer were created
-	exists, err := issuerExists(client, caSelfSignedIssuerName, localvz.Spec.Components.CertManager.Certificate.CA.ClusterResourceNamespace)
-	assert.NoError(t, err)
-	assert.True(t, exists)
-
-	exists, err = certificateExists(client, caCertificateName, localvz.Spec.Components.CertManager.Certificate.CA.ClusterResourceNamespace)
-	assert.NoError(t, err)
-	assert.True(t, exists)
-
-	exists, err = clusterIssuerExists(client, verrazzanoClusterIssuerName)
-	assert.NoError(t, err)
-	assert.True(t, exists)
-
-	// GIVEN that a secret with the cluster CA certificate exists
-	// WHEN a call is made to create the CA resources
-	// THEN the call succeeds and only a ClusterIssuer has been created
-	secret := v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      localvz.Spec.Components.CertManager.Certificate.CA.SecretName,
-			Namespace: localvz.Spec.Components.CertManager.Certificate.CA.ClusterResourceNamespace,
-		},
-	}
-	client = fake.NewClientBuilder().WithScheme(testScheme).WithObjects(&secret).Build()
-
-	opResult, err = createOrUpdateCAResources(spi.NewFakeContext(client, localvz, nil, false, profileDir))
-	assert.NoError(t, err)
-	assert.Equal(t, controllerutil.OperationResultCreated, opResult)
-
-	// validate that only the ClusterIssuer was created
-	exists, err = issuerExists(client, caSelfSignedIssuerName, localvz.Spec.Components.CertManager.Certificate.CA.ClusterResourceNamespace)
-	assert.NoError(t, err)
-	assert.False(t, exists)
-
-	exists, err = certificateExists(client, caCertificateName, localvz.Spec.Components.CertManager.Certificate.CA.ClusterResourceNamespace)
-	assert.NoError(t, err)
-	assert.False(t, exists)
-
-	exists, err = clusterIssuerExists(client, verrazzanoClusterIssuerName)
-	assert.NoError(t, err)
-	assert.True(t, exists)
-}
-
-// TestRenewAllCertificatesNoCertsPresent tests the checkRenewAllCertificates function (code coverage mainly)
-// GIVEN a call to checkRenewAllCertificates
-//
-//	WHEN No certs are found
-//	THEN no error is returned
-func TestRenewAllCertificatesNoCertsPresent(t *testing.T) {
-	client := fake.NewClientBuilder().WithScheme(testScheme).Build()
-	fakeContext := spi.NewFakeContext(client, defaultVZConfig, nil, false)
-	assert.NoError(t, checkRenewAllCertificates(fakeContext, true))
-}
-
-// TestDeleteObject tests the deleteObject function
-// GIVEN a call to deleteObject
-// WHEN for an object that exists
-// THEN no error is returned and the object is deleted, and that it is idempotent if called again for the same object
-func TestDeleteObject(t *testing.T) {
-	const name = "mysecret"
-	const ns = "myns"
-	secretToDelete := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{Name: name, Namespace: ns},
-	}
-	client := fake.NewClientBuilder().WithScheme(testScheme).
-		WithObjects(
-			secretToDelete,
-		).Build()
-
-	fakeContext := spi.NewFakeContext(client, &vzapi.Verrazzano{}, nil, false, profileDir)
-	assert.NoError(t, deleteObject(fakeContext.Client(), name, ns, &v1.Secret{}))
-	err := client.Get(context.TODO(), types.NamespacedName{Name: name, Namespace: ns}, &v1.Secret{})
-	assert.Error(t, err)
-	assert.True(t, errors.IsNotFound(err))
-
-	// Ensure this is idempotent, and no error is returned if the secret isn't found
-	assert.NoError(t, deleteObject(fakeContext.Client(), name, ns, &v1.Secret{}))
-
-}
-
-// TestCleanupUnusedACMEResources tests the cleanupUnusedResources function
-// GIVEN a call to cleanupUnusedResources
-// WHEN the default issuer is configured and there are leftover ACME resources
-// THEN no error is returned and default ACME CA secret is deleted without affecting the required default issuer resources
-func TestCleanupUnusedACMEResources(t *testing.T) {
-	vz := defaultVZConfig.DeepCopy()
-
-	client := fake.NewClientBuilder().WithScheme(testScheme).
-		WithObjects(createACMEResources()...).
-		WithObjects(createClusterIssuerResources()...).
-		WithObjects(createDefaultIssuerResources()...).
-		Build()
-
-	fakeContext := spi.NewFakeContext(client, vz, nil, false, profileDir)
-	assert.NoError(t, cleanupUnusedResources(fakeContext, true))
-	assertFound(t, client, verrazzanoClusterIssuerName, ComponentNamespace, &certv1.ClusterIssuer{})
-	assertNotFound(t, client, caAcmeSecretName, ComponentNamespace, &v1.Secret{})
-	assertFound(t, client, defaultCACertificateSecretName, ComponentNamespace, &v1.Secret{})
-	assertFound(t, client, caCertificateName, ComponentNamespace, &certv1.Certificate{})
-	assertFound(t, client, caSelfSignedIssuerName, ComponentNamespace, &certv1.Issuer{})
-}
-
-// TestCleanupUnusedDefaultCAResources tests the cleanupUnusedResources function
-// GIVEN a call to cleanupUnusedResources
-// WHEN the ACME issuer is configured and there are leftover default issuer resources
-// THEN no error is returned and default issuer resources are deleted without affecting the required ACME resources
-func TestCleanupUnusedDefaultCAResources(t *testing.T) {
-	vz := defaultVZConfig.DeepCopy()
-	vz.Spec.Components.CertManager.Certificate = vzapi.Certificate{}
-	vz.Spec.Components.CertManager.Certificate.Acme = acme
-
-	client := fake.NewClientBuilder().WithScheme(testScheme).
-		WithObjects(createACMEResources()...).
-		WithObjects(createClusterIssuerResources()...).
-		WithObjects(createDefaultIssuerResources()...).
-		Build()
-
-	fakeContext := spi.NewFakeContext(client, vz, nil, false, profileDir)
-	assert.NoError(t, cleanupUnusedResources(fakeContext, false))
-	assertFound(t, client, caAcmeSecretName, ComponentNamespace, &v1.Secret{})
-	assertFound(t, client, verrazzanoClusterIssuerName, ComponentNamespace, &certv1.ClusterIssuer{})
-	assertNotFound(t, client, defaultCACertificateSecretName, ComponentNamespace, &v1.Secret{})
-	assertNotFound(t, client, caCertificateName, ComponentNamespace, &certv1.Certificate{})
-	assertNotFound(t, client, caSelfSignedIssuerName, ComponentNamespace, &certv1.Issuer{})
-}
-
-// TestCustomCAConfigCleanupUnusedResources tests the cleanupUnusedResources function
-// GIVEN a call to cleanupUnusedResources
-// WHEN a Custom CA issuer is configured and there are leftover default and ACME issuer resources
-// THEN no error is returned and all leftover default and ACME issuer resources are deleted without affecting the required resources
-func TestCustomCAConfigCleanupUnusedResources(t *testing.T) {
-	const customCAName = "my-ca"
-	const customCANamespace = "customca"
-
-	vz := defaultVZConfig.DeepCopy()
-	vz.Spec.Components.CertManager.Certificate = vzapi.Certificate{}
-	vz.Spec.Components.CertManager.Certificate.CA = vzapi.CA{
-		SecretName:               customCAName,
-		ClusterResourceNamespace: customCANamespace,
-	}
-
-	client := fake.NewClientBuilder().WithScheme(testScheme).
-		WithObjects(createClusterIssuerResources()...).
-		WithObjects(createCustomCAResources(customCAName, customCANamespace)...).
-		WithObjects(createACMEResources()...).
-		WithObjects(createDefaultIssuerResources()...).
-		Build()
-
-	fakeContext := spi.NewFakeContext(client, vz, nil, false, profileDir)
-	assert.NoError(t, cleanupUnusedResources(fakeContext, true))
-	assertFound(t, client, customCAName, customCANamespace, &v1.Secret{})
-	assertFound(t, client, verrazzanoClusterIssuerName, ComponentNamespace, &certv1.ClusterIssuer{})
-	assertNotFound(t, client, caAcmeSecretName, ComponentNamespace, &v1.Secret{})
-	assertNotFound(t, client, defaultCACertificateSecretName, ComponentNamespace, &v1.Secret{})
-	assertNotFound(t, client, caCertificateName, ComponentNamespace, &certv1.Certificate{})
-	assertNotFound(t, client, caSelfSignedIssuerName, ComponentNamespace, &certv1.Issuer{})
-}
-
 // TestUninstallCertManager tests the cert-manager uninstall process
 // GIVEN a call to uninstallCertManager
 // WHEN the objects exist in the cluster
@@ -548,35 +382,6 @@ func TestUninstallCertManager(t *testing.T) {
 	certNS := v1.Namespace{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: constants.CertManagerNamespace,
-		},
-	}
-	vzCI := certv1.ClusterIssuer{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: verrazzanoClusterIssuerName,
-		},
-	}
-	ssIssuer := certv1.Issuer{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: ComponentNamespace,
-			Name:      caSelfSignedIssuerName,
-		},
-	}
-	caCert := certv1.Certificate{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: ComponentNamespace,
-			Name:      caCertificateName,
-		},
-	}
-	caSecret := v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: ComponentNamespace,
-			Name:      defaultCACertificateSecretName,
-		},
-	}
-	caAcmeSec := v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: ComponentNamespace,
-			Name:      caAcmeSecretName,
 		},
 	}
 
@@ -608,61 +413,6 @@ func TestUninstallCertManager(t *testing.T) {
 				&certNS,
 			},
 		},
-		{
-			name: "test cluster issuer",
-			objects: []clipkg.Object{
-				&controllerCM,
-				&caCM,
-				&certNS,
-				&vzCI,
-			},
-		},
-		{
-			name: "test issuer",
-			objects: []clipkg.Object{
-				&controllerCM,
-				&caCM,
-				&certNS,
-				&vzCI,
-				&ssIssuer,
-			},
-		},
-		{
-			name: "test ca cert",
-			objects: []clipkg.Object{
-				&controllerCM,
-				&caCM,
-				&certNS,
-				&vzCI,
-				&ssIssuer,
-				&caCert,
-			},
-		},
-		{
-			name: "test ca secret",
-			objects: []clipkg.Object{
-				&controllerCM,
-				&caCM,
-				&certNS,
-				&vzCI,
-				&ssIssuer,
-				&caCert,
-				&caSecret,
-			},
-		},
-		{
-			name: "test ca acme secret",
-			objects: []clipkg.Object{
-				&controllerCM,
-				&caCM,
-				&certNS,
-				&vzCI,
-				&ssIssuer,
-				&caCert,
-				&caSecret,
-				&caAcmeSec,
-			},
-		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -676,21 +426,6 @@ func TestUninstallCertManager(t *testing.T) {
 			// expect the CA injector ConfigMap to get deleted
 			err = c.Get(context.TODO(), types.NamespacedName{Name: caInjectorConfigMap, Namespace: constants.KubeSystem}, &v1.ConfigMap{})
 			assert.Error(t, err, "Expected the ConfigMap %s to be deleted", caInjectorConfigMap)
-			// expect the Namespace to get deleted
-			err = c.Get(context.TODO(), types.NamespacedName{Name: verrazzanoClusterIssuerName, Namespace: vzconst.DefaultNamespace}, &certv1.ClusterIssuer{})
-			assert.Error(t, err, "Expected the ClusterIssuer %s to be deleted", ComponentNamespace)
-			// expect the Namespace to get deleted
-			err = c.Get(context.TODO(), types.NamespacedName{Name: caSelfSignedIssuerName, Namespace: ComponentNamespace}, &certv1.Issuer{})
-			assert.Error(t, err, "Expected the Issuer %s to be deleted", ComponentNamespace)
-			// expect the Namespace to get deleted
-			err = c.Get(context.TODO(), types.NamespacedName{Name: caCertificateName, Namespace: ComponentNamespace}, &certv1.Certificate{})
-			assert.Error(t, err, "Expected the Certificate %s to be deleted", ComponentNamespace)
-			// expect the Namespace to get deleted
-			err = c.Get(context.TODO(), types.NamespacedName{Name: defaultCACertificateSecretName, Namespace: ComponentNamespace}, &v1.Secret{})
-			assert.Error(t, err, "Expected the Secret %s to be deleted", ComponentNamespace)
-			// expect the Namespace to get deleted
-			err = c.Get(context.TODO(), types.NamespacedName{Name: caAcmeSecretName, Namespace: ComponentNamespace}, &v1.Secret{})
-			assert.Error(t, err, "Expected the Secret %s to be deleted", ComponentNamespace)
 		})
 	}
 }
@@ -764,7 +499,7 @@ func TestGetOverrides(t *testing.T) {
 // WHEN the CR passed is v1alpha1
 // THEN it is inspected to validate the host name length of endpoints
 func TestValidateLongestHostName(t *testing.T) {
-	assert := assert.New(t)
+	asserts := assert.New(t)
 	cr1, cr2, cr3, cr4, cr5, cr6 := *vz.DeepCopy(), *vz.DeepCopy(), *vz.DeepCopy(), *vz.DeepCopy(), *vz.DeepCopy(), *vz.DeepCopy()
 	cr1.Spec.Components.DNS.OCI = ociLongDNSZoneName
 	cr2.Spec.Components.DNS.OCI = oci
@@ -809,9 +544,9 @@ func TestValidateLongestHostName(t *testing.T) {
 	for _, test := range tests {
 		err := validateLongestHostName(&test.cr)
 		if test.wantError {
-			assert.EqualError(err, test.want)
+			asserts.EqualError(err, test.want)
 		} else {
-			assert.NoError(err)
+			asserts.NoError(err)
 		}
 	}
 }
@@ -821,7 +556,7 @@ func TestValidateLongestHostName(t *testing.T) {
 // WHEN the CR passed is v1beta1
 // THEN it is inspected to validate the host name length of endpoints
 func TestValidateLongestHostNameV1Beta1(t *testing.T) {
-	assert := assert.New(t)
+	asserts := assert.New(t)
 	cr1, cr2, cr3, cr4, cr5, cr6 := *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy()
 	cr1.Spec.Components.DNS.OCI = ociLongDNSZoneNameV1Beta1
 	cr2.Spec.Components.DNS.OCI = ociV1Beta1
@@ -866,83 +601,11 @@ func TestValidateLongestHostNameV1Beta1(t *testing.T) {
 	for _, test := range tests {
 		err := validateLongestHostName(&test.cr)
 		if test.wantError {
-			assert.EqualError(err, test.want)
+			asserts.EqualError(err, test.want)
 		} else {
-			assert.NoError(err)
+			asserts.NoError(err)
 		}
 	}
-}
-
-func assertNotFound(t *testing.T, client clipkg.WithWatch, name string, namespace string, obj clipkg.Object) {
-	err := client.Get(context.TODO(), types.NamespacedName{Name: name, Namespace: namespace}, obj)
-	assert.Error(t, err)
-	assert.True(t, errors.IsNotFound(err))
-}
-
-func assertFound(t *testing.T, client clipkg.WithWatch, name string, namespace string, obj clipkg.Object) {
-	err := client.Get(context.TODO(), types.NamespacedName{Name: name, Namespace: namespace}, obj)
-	assert.NoError(t, err)
-}
-
-func createACMEResources() []clipkg.Object {
-	acmeSecret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{Name: caAcmeSecretName, Namespace: ComponentNamespace},
-	}
-	return []clipkg.Object{acmeSecret}
-}
-
-func createCustomCAResources(name string, namespace string) []clipkg.Object {
-	customcA := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{Name: name, Namespace: namespace},
-	}
-	return []clipkg.Object{customcA}
-}
-
-func createClusterIssuerResources() []clipkg.Object {
-	customcA := &certv1.ClusterIssuer{
-		ObjectMeta: metav1.ObjectMeta{Name: verrazzanoClusterIssuerName, Namespace: ComponentNamespace},
-	}
-	return []clipkg.Object{customcA}
-}
-
-func createDefaultIssuerResources() []clipkg.Object {
-	secretToDelete := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{Name: defaultCACertificateSecretName, Namespace: ComponentNamespace},
-	}
-	selfSignedCACert := &certv1.Certificate{
-		ObjectMeta: metav1.ObjectMeta{Name: caCertificateName, Namespace: ComponentNamespace},
-	}
-	selfSignedIssuer := &certv1.Issuer{
-		ObjectMeta: metav1.ObjectMeta{Name: caSelfSignedIssuerName, Namespace: ComponentNamespace},
-	}
-	return []clipkg.Object{secretToDelete, selfSignedCACert, selfSignedIssuer}
-}
-
-// issuerExists returns true if the Issuer with the name and namespace exists.
-func issuerExists(client clipkg.Client, name string, namespace string) (bool, error) {
-	issuer := certv1.Issuer{}
-	if err := client.Get(context.TODO(), clipkg.ObjectKey{Name: name, Namespace: namespace}, &issuer); err != nil {
-		return false, clipkg.IgnoreNotFound(err)
-	}
-	return true, nil
-}
-
-// clusterIssuerExists returns true if the ClusterIssuer with the name exists.
-func clusterIssuerExists(client clipkg.Client, name string) (bool, error) {
-	clusterIssuer := certv1.ClusterIssuer{}
-	if err := client.Get(context.TODO(), clipkg.ObjectKey{Name: name}, &clusterIssuer); err != nil {
-		return false, clipkg.IgnoreNotFound(err)
-	}
-	return true, nil
-}
-
-// certificateExists returns true if the Certificate with the name and namespace exists.
-func certificateExists(client clipkg.Client, name string, namespace string) (bool, error) {
-	cert := certv1.Certificate{}
-	if err := client.Get(context.TODO(), clipkg.ObjectKey{Name: name, Namespace: namespace}, &cert); err != nil {
-		return false, clipkg.IgnoreNotFound(err)
-	}
-	return true, nil
 }
 
 // Create a new deployment object for testing
@@ -999,6 +662,88 @@ func newReplicaSet(name string) *appsv1.ReplicaSet {
 			Annotations: map[string]string{"deployment.kubernetes.io/revision": "1"},
 		},
 	}
+}
+
+func createCertSecretNoParent(name string, namespace string, cn string) (*v1.Secret, error) {
+	fakeCertBytes, err := createFakeCertBytes(cn, nil)
+	if err != nil {
+		return nil, err
+	}
+	secret := &v1.Secret{
+		TypeMeta: metav1.TypeMeta{},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+		Data: map[string][]byte{
+			v1.TLSCertKey: fakeCertBytes,
+		},
+		Type: v1.SecretTypeTLS,
+	}
+	return secret, nil
+}
+
+var testRSAKey *rsa.PrivateKey
+
+func getRSAKey() (*rsa.PrivateKey, error) {
+	if testRSAKey == nil {
+		var err error
+		if testRSAKey, err = rsa.GenerateKey(rand.Reader, 2048); err != nil {
+			return nil, err
+		}
+	}
+	return testRSAKey, nil
+}
+
+func createFakeCertBytes(cn string, parent *x509.Certificate) ([]byte, error) {
+	rsaKey, err := getRSAKey()
+	if err != nil {
+		return []byte{}, err
+	}
+
+	cert := createFakeCertificate(cn)
+	if parent == nil {
+		parent = cert
+	}
+	pubKey := &rsaKey.PublicKey
+	certBytes, err := x509.CreateCertificate(rand.Reader, cert, parent, pubKey, rsaKey)
+	if err != nil {
+		return []byte{}, err
+	}
+	certPem := pem.EncodeToMemory(&pem.Block{
+		Type:  "CERTIFICATE",
+		Bytes: certBytes,
+	})
+	return certPem, nil
+}
+
+func createFakeCertificate(cn string) *x509.Certificate {
+	duration30, _ := time.ParseDuration("-30h")
+	notBefore := time.Now().Add(duration30) // valid 30 hours ago
+	duration1Year, _ := time.ParseDuration("90h")
+	notAfter := notBefore.Add(duration1Year) // for 90 hours
+	serialNo := big.NewInt(int64(123123413123))
+	cert := &x509.Certificate{
+		Subject: pkix.Name{
+			Country:      []string{"US"},
+			Organization: []string{"BarOrg"},
+			SerialNumber: "2234",
+			CommonName:   cn,
+		},
+		SerialNumber:          serialNo,
+		NotBefore:             notBefore,
+		NotAfter:              notAfter,
+		KeyUsage:              x509.KeyUsageKeyEncipherment,
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
+		BasicConstraintsValid: true,
+		IsCA:                  false,
+		SignatureAlgorithm:    x509.SHA256WithRSA,
+	}
+
+	cert.IPAddresses = append(cert.IPAddresses, net.ParseIP("127.0.0.1"))
+	cert.IPAddresses = append(cert.IPAddresses, net.ParseIP("::"))
+	cert.DNSNames = append(cert.DNSNames, "localhost")
+	return cert
 }
 
 // Create a bool pointer
