+++ platform-operator/controllers/verrazzano/component/registry/registry.go
@@ -10,6 +10,7 @@ import (
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/authproxy"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/capi"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanagerconfig"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/clusteroperator"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/coherence"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/console"
@@ -72,6 +73,7 @@ func InitRegistry() {
 		weblogic.NewComponent(),
 		nginx.NewComponent(),
 		certmanager.NewComponent(),
+		certmanagerconfig.NewComponent(),
 		externaldns.NewComponent(),
 		capi.NewComponent(),
 		rancher.NewComponent(),
