+++ platform-operator/controllers/verrazzano/component/thanos/thanos_test.go
@@ -133,6 +133,8 @@ func TestAppendOverrides(t *testing.T) {
 		"queryFrontend.ingress.extraTls[0].hosts[0]":                                           "thanos-query.default.11.22.33.44.nip.io",
 		"queryFrontend.ingress.extraTls[0].secretName":                                         queryCertificateName,
 		`queryFrontend.ingress.annotations.nginx\.ingress\.kubernetes\.io/session-cookie-name`: queryHostName,
+		`queryFrontend.ingress.annotations.cert-manager\.io/cluster-issuer`:                    globalconst.VerrazzanoClusterIssuerName,
+		`queryFrontend.ingress.annotations.cert-manager\.io/common-name`:                       "thanos-query.default.11.22.33.44.nip.io",
 		"query.ingress.grpc.namespace":                                                         constants.VerrazzanoSystemNamespace,
 		"query.ingress.grpc.ingressClassName":                                                  "verrazzano-nginx",
 		"query.ingress.grpc.extraRules[0].host":                                                "thanos-query-store.default.11.22.33.44.nip.io",
@@ -142,12 +144,16 @@ func TestAppendOverrides(t *testing.T) {
 		"query.ingress.grpc.extraRules[0].http.paths[0].pathType":                              string(netv1.PathTypeImplementationSpecific),
 		"query.ingress.grpc.extraTls[0].hosts[0]":                                              "thanos-query-store.default.11.22.33.44.nip.io",
 		"query.ingress.grpc.extraTls[0].secretName":                                            queryStoreCertificateName,
+		`query.ingress.grpc.annotations.cert-manager\.io/cluster-issuer`:                       globalconst.VerrazzanoClusterIssuerName,
+		`query.ingress.grpc.annotations.cert-manager\.io/common-name`:                          "thanos-query-store.default.11.22.33.44.nip.io",
 	}
 	sslipioKVs := map[string]string{
-		"queryFrontend.ingress.extraRules[0].host":   "thanos-query.default.11.22.33.44.sslip.io",
-		"queryFrontend.ingress.extraTls[0].hosts[0]": "thanos-query.default.11.22.33.44.sslip.io",
-		"query.ingress.grpc.extraRules[0].host":      "thanos-query-store.default.11.22.33.44.sslip.io",
-		"query.ingress.grpc.extraTls[0].hosts[0]":    "thanos-query-store.default.11.22.33.44.sslip.io",
+		"queryFrontend.ingress.extraRules[0].host":                       "thanos-query.default.11.22.33.44.sslip.io",
+		"queryFrontend.ingress.extraTls[0].hosts[0]":                     "thanos-query.default.11.22.33.44.sslip.io",
+		`queryFrontend.ingress.annotations.cert-manager\.io/common-name`: "thanos-query.default.11.22.33.44.sslip.io",
+		"query.ingress.grpc.extraRules[0].host":                          "thanos-query-store.default.11.22.33.44.sslip.io",
+		"query.ingress.grpc.extraTls[0].hosts[0]":                        "thanos-query-store.default.11.22.33.44.sslip.io",
+		`query.ingress.grpc.annotations.cert-manager\.io/common-name`:    "thanos-query-store.default.11.22.33.44.sslip.io",
 	}
 	istioEnabledKV := map[string]string{
 		"verrazzano.isIstioEnabled": "true",
@@ -162,10 +168,12 @@ func TestAppendOverrides(t *testing.T) {
 		"queryFrontend.ingress.extraTls[0].hosts[0]":                                   "thanos-query.default.mydomain.com",
 		`queryFrontend.ingress.annotations.external-dns\.alpha\.kubernetes\.io/target`: "verrazzano-ingress.default.mydomain.com",
 		`queryFrontend.ingress.annotations.external-dns\.alpha\.kubernetes\.io/ttl`:    "60",
+		`queryFrontend.ingress.annotations.cert-manager\.io/common-name`:               "thanos-query.default.mydomain.com",
 		"query.ingress.grpc.extraRules[0].host":                                        "thanos-query-store.default.mydomain.com",
 		"query.ingress.grpc.extraTls[0].hosts[0]":                                      "thanos-query-store.default.mydomain.com",
 		`query.ingress.grpc.annotations.external-dns\.alpha\.kubernetes\.io/target`:    "verrazzano-ingress.default.mydomain.com",
 		`query.ingress.grpc.annotations.external-dns\.alpha\.kubernetes\.io/ttl`:       "60",
+		`query.ingress.grpc.annotations.cert-manager\.io/common-name`:                  "thanos-query-store.default.mydomain.com",
 	}
 	externalDNSZone := "mydomain.com"
